requirements.txt
gdown==5.2.0
numpy==2.1.3
Pillow==10.1.0
Pillow==11.0.0
streamlit==1.40.0
tensorflow==2.18.0
